requirements.txt
cryptography
Django
djangorestframework
ecdsa
pytz
